feat(Client sendModel): send model once in the end of constructor
This way we are sure that we are sending model only once.
If we write something into model after bundling and before starting it
in browser, racer should catch up in the browser to the right state of
model. So it should be alright to use model even if it is not yet sent
to the browser.