feat(Workspace): fetch model in the constructor
If there are call to updateModel() before model was fetched, these calls
are queued and model is actually updated after it was fetched.
This way we only fetch model once inside constructor and racer is
deciding when to write into the model.

This change is similar to the change in Client.
Client change hash - d7ddb5e4cbcb08db95b13b2a65fa71c9518f936d