fix(Workspace): updateModel() in the end of constructor
Previous commit with cleaning model before exiting fixed problem with
updating after fetch. Now this works similar to the Client module.

We updateModel in the end of constructor while model is not fetched
yet. We write into model anyway and racer will decide for itself when
it is going to fetch and actually update model in database.