fix(Server DEFAULT_OPTS): change default path to static folder
Because we didnt resolve path in the beginning, it was trying to find
folder 'public' in the folder with current file that we are running.
Now, we have __dirname. It resolves to the folder of the current file
'Server.js' and so we are looking for folder 'public' in folder of the
server file.