feat(racer): pass only store and recreate model for each class
Previously we created one store and one model in index.js file and
passed both of these objects between different classes. This way we had
only one model always. We also had problems when we were trying to
subscribe to the same model few times.
Now we create one store in index.js file and pass it to different
classes and we create new model from that store object for each
particular class. This way classes are more independent and have more
control over their own instance of Model.

We also do not sent mainPath anymore as it exists in util module and we
can get it from there.