feat(WorkspaceObject): fetch model in constructor
If there are calls to updateModel() before model is fetched, these calls
are queued and model is actually updated after it was fetched.
This way we only fetch model once inside constructor and writing into
it is handled by racer itself.

This change is similar to the change in Client and Workspace.
   Client change hash - d7ddb5e4cbcb08db95b13b2a65fa71c9518f936d
   Workspace change hash - 459693e99a95747fb08ba6230cbb20b5a9f661de