feat(index): set to empty instead of deleting while cleaning model
Because of the move to new way with few models recreated for each class
instead of one model that was shared between everyone, we faced
issue #1. Every time when we delete collections on exit, we need to
recreate them on new start, and because of that somewhere inside sharejs
there are two events (that create doc) happen. It seems that sharejs
cannot resolve 2 create events, and it crashes with error 'Document
created remotely'. So to eliminate creating new collections each time,
we are just emptying them on exit. We still have the problem when we
start application for the very first time and database is absolutely
empty.